+++ app/src/main/java/com/ft/webview/TCBaseWebView.java
@@ -1,0 +1,36 @@
+package com.ft.webview;
+
+import android.content.Context;
+import android.util.AttributeSet;
+import android.webkit.WebView;
+
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+
+
+public class TCBaseWebView extends WebView {
+    public TCBaseWebView(@NonNull Context context) {
+        super(context);
+    }
+
+    public TCBaseWebView(@NonNull Context context, @Nullable AttributeSet attrs) {
+        super(context, attrs);
+    }
+
+    public TCBaseWebView(@NonNull Context context, @Nullable AttributeSet attrs, int defStyleAttr) {
+        super(context, attrs, defStyleAttr);
+    }
+
+    public TCBaseWebView(@NonNull Context context, @Nullable AttributeSet attrs, int defStyleAttr, int defStyleRes) {
+        super(context, attrs, defStyleAttr, defStyleRes);
+    }
+
+    public TCBaseWebView(@NonNull Context context, @Nullable AttributeSet attrs, int defStyleAttr, boolean privateBrowsing) {
+        super(context, attrs, defStyleAttr, privateBrowsing);
+    }
+
+    @Override
+    public void loadUrl(String s) {
+        TCBaseWebView.super.loadUrl(s);
+    }
+}